# docs/openapi.yaml
openapi: 3.1.0
info:
  title: REKAM API
  version: 0.1.0
  description: Backend API for REKAM local and on-premise deployment.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/users/{id}:
    put:
      summary: Update user
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: User updated. }
  /api/users/{id}/reset-password:
    put:
      summary: Reset user password
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Password reset. }
...
  /api/media/upload:
    post:
      summary: Upload media files
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
      responses:
        "201": { description: Media uploaded. }
...
  /p/{token}/media:
    get:
      security: []
      summary: Public QR media list
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Public media. }
...